GestionPersonnel/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
PyJWT==2.8.0
python-decouple==3.8
python-environ==0.4.54
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
